% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcArrowGrob.R
\name{arcArrowGrob}
\alias{arcArrowGrob}
\title{Create an arc arrow grob}
\usage{
arcArrowGrob(
  xmin = NULL,
  xmax = NULL,
  y = NULL,
  width = 0.5,
  start.arrow.len = 6,
  end.arrow.len = 6,
  arrow.start.width = 0.5,
  arrow.end.width = 0.5,
  start = 0,
  end = 360,
  r0 = 0.5,
  r1 = 1,
  x0 = 0,
  y0 = 0,
  polygon.gp = NULL,
  n = 100,
  clock.wise = FALSE,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  xscale = NULL,
  yscale = NULL,
  ...,
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{xmin}{Minimum x-coordinate for the arrow.}

\item{xmax}{Maximum x-coordinate for the arrow.}

\item{y}{Y-coordinate for the arrow.}

\item{width}{Width of the arrow.}

\item{start.arrow.len}{Length of the starting arrow.}

\item{end.arrow.len}{Length of the ending arrow.}

\item{arrow.start.width}{Width of the starting arrow.}

\item{arrow.end.width}{Width of the ending arrow.}

\item{start}{Starting angle for the arc (in degrees).}

\item{end}{Ending angle for the arc (in degrees).}

\item{r0}{Inner radius of the arc.}

\item{r1}{Outer radius of the arc.}

\item{x0}{X-coordinate of the center of the arc.}

\item{y0}{Y-coordinate of the center of the arc.}

\item{polygon.gp}{Graphic parameters for the arc.}

\item{n}{Number of points to use for drawing the arc.}

\item{clock.wise}{If TRUE, draw the arc in a clockwise direction.}

\item{extend.xscale}{Scale factor for extending the x-axis range.}

\item{extend.yscale}{Scale factor for extending the y-axis range.}

\item{xscale}{Optional custom x-scale.}

\item{yscale}{Optional custom y-scale.}

\item{...}{Additional arguments (not used).}

\item{name}{Name for the grob.}

\item{gp}{Graphic parameters for the grob.}

\item{vp}{Viewport for the grob.}
}
\value{
An arc arrow grob.
}
\description{
This function creates a graphical object (grob) representing an arrow-shaped arc.
}
