% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcLinesGrob.R
\name{arcLinesGrob}
\alias{arcLinesGrob}
\title{Create a custom arc lines graphical object.}
\usage{
arcLinesGrob(
  x = NULL,
  y = NULL,
  start = 0,
  end = 60,
  r0 = 0.5,
  r1 = 1,
  x0 = 0,
  y0 = 0,
  lines.gp = NULL,
  arrow = NULL,
  id = NULL,
  n = 100,
  clock.wise = FALSE,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  xscale = NULL,
  yscale = NULL,
  ...,
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{x}{The numeric values at which the point should be placed along the arc.
Default is NULL.}

\item{y}{The numeric values at which the point should be placed along the arc.
Default is NULL.}

\item{start}{The starting angle of the arc (in degrees). Default is 0.}

\item{end}{The ending angle of the arc (in degrees). Default is 60.}

\item{r0}{The inner radius of the arc. Default is 0.5.}

\item{r1}{The outer radius of the arc. Default is 1.}

\item{x0}{The x-coordinate of the center of the unit circle. Default is 0.}

\item{y0}{The y-coordinate of the center of the unit circle. Default is 0.}

\item{lines.gp}{A graphical parameter object for the lines' appearance. Default is NULL.}

\item{arrow}{A list specifying arrow settings for the lines. Default is NULL.}

\item{id}{An identifier for each line segment. Default is NULL.}

\item{n}{The number of points used to approximate the lines. Default is 100.}

\item{clock.wise}{A logical value indicating whether the arc is drawn in a
clockwise direction. Default is FALSE.}

\item{extend.xscale}{The extension factor for the scaling of the arc. Default is 0.05.}

\item{extend.yscale}{The extension factor for the scaling of the arc. Default is 0.05.}

\item{xscale}{The scaling range for the x-axis. Default is NULL.}

\item{yscale}{The scaling range for the y-axis. Default is NULL.}

\item{...}{Additional arguments to be passed to the \code{polylineGrob} function.}

\item{name}{The name of the arc lines graphical object. Default is NULL.}

\item{gp}{A graphical parameter object for the arc lines. Default is NULL.}

\item{vp}{A viewport object for the arc lines. Default is NULL.}
}
\value{
A graphical object of class 'grob'.
}
\description{
This function creates a custom arc lines graphical object that can be added to
a plot in R. The lines are defined by their positions along an arc in a polar
coordinate system.
}
\examples{
\dontrun{
# Create arc lines
arc_lines <- arcLinesGrob(x = c(0, 1, 2), y = c(0.5, 0.7, 0.9),
                           start = 0, end = 180, r0 = 0.5, r1 = 1)
# Add the arc lines to a plot
grid.draw(arc_lines)
}

}
