% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcLinks2Grob.R
\name{arcLinks2Grob}
\alias{arcLinks2Grob}
\title{Create a Grob for Arc Links}
\usage{
arcLinks2Grob(
  start = 0,
  end = 60,
  r0 = 1,
  r1 = 1,
  r = 1,
  arrow.len = 0.05,
  x0 = 0,
  y0 = 0,
  curve.arrow = NULL,
  curve.height = 0.5,
  start.arrow = FALSE,
  end.arrow = FALSE,
  bezierCurve.gp = gpar(fill = "black"),
  bezierPolygon.gp = gpar(fill = "black", col = "black"),
  n = 100,
  clock.wise = FALSE,
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{start}{Starting angle in degrees (default: 0).}

\item{end}{Ending angle in degrees (default: 60).}

\item{r0}{Inner radius (default: 1).}

\item{r1}{Outer radius (default: 1).}

\item{r}{Radius for the arc (default: 1).}

\item{arrow.len}{Length of arrows (default: 0.05).}

\item{x0}{X-coordinate of the center point (default: 0).}

\item{y0}{Y-coordinate of the center point (default: 0).}

\item{curve.arrow}{Arrow for the curve (default: NULL).}

\item{curve.height}{Height of the curve (default: 0.5).}

\item{start.arrow}{Include arrow at the starting point (default: FALSE).}

\item{end.arrow}{Include arrow at the ending point (default: FALSE).}

\item{bezierCurve.gp}{Graphics parameters for the curve (default: gpar(fill = "black")).}

\item{bezierPolygon.gp}{Graphics parameters for the polygon (default: gpar(fill = "black", col = "black")).}

\item{n}{Number of points for the curve (default: 100).}

\item{clock.wise}{Direction of the curve (default: FALSE).}

\item{name}{Name of the grob (default: NULL).}

\item{gp}{Graphics parameters for the grob (default: NULL).}

\item{vp}{Viewport for the grob (default: NULL).}
}
\value{
A grob representing the arc link.
}
\description{
This function creates a grob for arc links between two points.
}
