% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcLinksGrob.R
\name{arcLinksGrob}
\alias{arcLinksGrob}
\title{Create custom arc links graphical object.}
\usage{
arcLinksGrob(
  start = 0,
  end = 60,
  r = 1,
  arrow.len = 0.05,
  x0 = 0,
  y0 = 0,
  curve.arrow = NULL,
  curve.height = 0.5,
  start.arrow = FALSE,
  end.arrow = FALSE,
  bezierCurve.gp = gpar(fill = "black"),
  bezierPolygon.gp = gpar(fill = "black", col = "black"),
  n = 100,
  clock.wise = FALSE,
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{start}{The starting angle of the arc (in degrees). Default is 0.}

\item{end}{The ending angle of the arc (in degrees). Default is 60.}

\item{r}{The radius of the arcs. Default is 1.}

\item{arrow.len}{The length of arrowheads on the links. Default is 0.05.}

\item{x0}{The x-coordinate of the center of the unit circle. Default is 0.}

\item{y0}{The y-coordinate of the center of the unit circle. Default is 0.}

\item{curve.arrow}{A graphical parameter object for the arrowheads on the curves.
Default is NULL.}

\item{curve.height}{The height of the control points for the Bezier curves.
Default is 0.5.}

\item{start.arrow}{A logical value indicating whether to add an arrowhead at
the starting point. Default is FALSE.}

\item{end.arrow}{A logical value indicating whether to add an arrowhead at
the ending point. Default is FALSE.}

\item{bezierCurve.gp}{A graphical parameter object for the Bezier curve appearance.
Default is gpar(fill = "black").}

\item{bezierPolygon.gp}{A graphical parameter object for the Bezier curve
polygon appearance. Default is gpar(fill = "black", col = "black").}

\item{n}{The number of points to generate along the arcs. Default is 100.}

\item{clock.wise}{A logical value indicating whether the arcs are drawn in a
clockwise direction. Default is FALSE.}

\item{name}{The name of the arc links graphical object. Default is NULL.}

\item{gp}{A graphical parameter object for the arc links. Default is NULL.}

\item{vp}{A viewport object for the arc links. Default is NULL.}
}
\value{
A graphical object of class 'grob'.
}
\description{
This function creates a custom arc links graphical object that can be added to
a plot in R. The links consist of Bezier curves and arcs connecting points in a polar
coordinate system.
}
\examples{
\dontrun{
# Create arc links
arc_links <- arcLinksGrob(start = 0, end = 120, r = 1, arrow.len = 0.05)
# Add the arc links to a plot
grid.draw(arc_links)
}

}
