% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcPointGrob.R
\name{arcPointGrob}
\alias{arcPointGrob}
\title{Create a custom arc point graphical object.}
\usage{
arcPointGrob(
  x = NULL,
  y = NULL,
  start = 0,
  end = 60,
  r0 = 0.5,
  r1 = 1,
  x0 = 0,
  y0 = 0,
  point.gp = NULL,
  size = 5,
  pch = 19,
  clock.wise = FALSE,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  xscale = NULL,
  yscale = NULL,
  ...,
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{x}{The numeric values at which the point should be placed along the arc.
Default is NULL.}

\item{y}{The numeric values at which the point should be placed along the arc.
Default is NULL.}

\item{start}{The starting angle of the arc (in degrees). Default is 0.}

\item{end}{The ending angle of the arc (in degrees). Default is 60.}

\item{r0}{The inner radius of the arc. Default is 0.5.}

\item{r1}{The outer radius of the arc. Default is 1.}

\item{x0}{The x-coordinate of the center of the unit circle. Default is 0.}

\item{y0}{The y-coordinate of the center of the unit circle. Default is 0.}

\item{point.gp}{A graphical parameter object for the point's appearance. Default is NULL.}

\item{size}{The size of the point. Default is 5.}

\item{pch}{The plotting character used for the point. Default is 19.}

\item{clock.wise}{A logical value indicating whether the arc is drawn in a
clockwise direction. Default is FALSE.}

\item{extend.xscale}{The extension factor for the scaling of the arc. Default is 0.05.}

\item{extend.yscale}{The extension factor for the scaling of the arc. Default is 0.05.}

\item{xscale}{The scaling range for the x-axis. Default is NULL.}

\item{yscale}{The scaling range for the y-axis. Default is NULL.}

\item{...}{Additional arguments to be passed to the \code{pointsGrob} function.}

\item{name}{The name of the arc point graphical object. Default is NULL.}

\item{gp}{A graphical parameter object for the arc point. Default is NULL.}

\item{vp}{A viewport object for the arc point. Default is NULL.}
}
\value{
A graphical object of class 'grob'.
}
\description{
This function creates a custom arc point graphical object that can be added to
a plot in R. The point's position is determined by its radius and angle in a
polar coordinate system.
}
\examples{
\dontrun{
# Create an arc point
arc_point <- arcPointGrob(x = 30, y = 0.7, start = 0, end = 180, r0 = 0.5, r1 = 1)
# Add the arc point to a plot
grid.draw(arc_point)
}

}
