% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcPolygonGrob.R
\name{arcPolygonGrob}
\alias{arcPolygonGrob}
\title{Create custom arc polygon graphical object.}
\usage{
arcPolygonGrob(
  x = NULL,
  y = NULL,
  start = 0,
  end = 60,
  r0 = 0.5,
  r1 = 1,
  x0 = 0,
  y0 = 0,
  polygon.gp = NULL,
  n = 100,
  id = NULL,
  polar.every = TRUE,
  clock.wise = FALSE,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  xscale = NULL,
  yscale = NULL,
  ...,
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{x}{The x-coordinates of the polygon points. Default is NULL.}

\item{y}{The y-coordinates of the polygon points. Default is NULL.}

\item{start}{The starting angle of the arc (in degrees). Default is 0.}

\item{end}{The ending angle of the arc (in degrees). Default is 60.}

\item{r0}{The inner radius of the arc. Default is 0.5.}

\item{r1}{The outer radius of the arc. Default is 1.}

\item{x0}{The x-coordinate of the center of the unit circle. Default is 0.}

\item{y0}{The y-coordinate of the center of the unit circle. Default is 0.}

\item{polygon.gp}{A graphical parameter object for the polygon's appearance.
Default is NULL.}

\item{n}{The number of points to generate along the arc. Default is 100.}

\item{id}{An optional vector specifying the ID or group for each point. Default is NULL.}

\item{polar.every}{A logical value indicating whether to calculate points for every
section of the arc. If FALSE, points will be calculated for the entire arc. Default is TRUE.}

\item{clock.wise}{A logical value indicating whether the arc is drawn in a clockwise
direction. Default is FALSE.}

\item{extend.xscale}{The extension factor for the scaling of the arc. Default is 0.05.}

\item{extend.yscale}{The extension factor for the scaling of the arc. Default is 0.05.}

\item{xscale}{The scaling range for the x-axis. Default is NULL.}

\item{yscale}{The scaling range for the y-axis. Default is NULL.}

\item{...}{unused params.}

\item{name}{The name of the arc polygon graphical object. Default is NULL.}

\item{gp}{A graphical parameter object for the arc polygon. Default is NULL.}

\item{vp}{A viewport object for the arc polygon. Default is NULL.}
}
\value{
A graphical object of class 'grob'.
}
\description{
This function creates a custom arc polygon graphical object that can be added to
a plot in R. The polygon is defined by connecting points along an arc in a polar
coordinate system.
}
\examples{
\dontrun{
# Create arc polygon
arc_polygon <- arcPolygonGrob(x = c(0, 10, 20), y = c(0, 0, 0))
# Add the arc polygon to a plot
grid.draw(arc_polygon)
}

}
