% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcSectorGrob.R
\name{arcSectorGrob}
\alias{arcSectorGrob}
\title{Create a custom arc sector graphical object.}
\usage{
arcSectorGrob(
  start = 0,
  end = 360,
  r0 = 0,
  r1 = 1,
  x0 = 0,
  y0 = 0,
  sector.gp = gpar(fill = "white"),
  n = 100,
  clock.wise = FALSE,
  add.xaxis = TRUE,
  add.yaxis = TRUE,
  arcxAxisGrob.params = list(),
  arcyAxisGrob.params = list(),
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  xaxis.extend.xscale = NULL,
  xaxis.extend.yscale = NULL,
  yaxis.extend.xscale = NULL,
  yaxis.extend.yscale = NULL,
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{start}{The starting angle of the arc sector (in degrees). Default is 0.}

\item{end}{The ending angle of the arc sector (in degrees). Default is 360.}

\item{r0}{The inner radius of the arc sector. Default is 0.}

\item{r1}{The outer radius of the arc sector. Default is 1.}

\item{x0}{The x-coordinate of the center of the unit circle. Default is 0.}

\item{y0}{The y-coordinate of the center of the unit circle. Default is 0.}

\item{sector.gp}{A graphical parameter object for the arc sector's appearance.
Default is gpar(fill = "white").}

\item{n}{The number of points used to approximate the arc sector. Default is 100.}

\item{clock.wise}{A logical value indicating whether the arc sector should be drawn
in a clockwise direction. Default is FALSE.}

\item{add.xaxis}{A logical value indicating whether to add an x-axis to the plot.
Default is TRUE.}

\item{add.yaxis}{A logical value indicating whether to add a y-axis to the plot.
Default is TRUE.}

\item{arcxAxisGrob.params}{A list of additional parameters to be passed to the
arcxAxisGrob function when adding the x-axis. Default is an empty list.}

\item{arcyAxisGrob.params}{A list of additional parameters to be passed to the
arcyAxisGrob function when adding the y-axis. Default is an empty list.}

\item{extend.xscale}{The extension factor for the scaling of the arc sector.
Default is 0.05.}

\item{extend.yscale}{The extension factor for the scaling of the arc sector.
Default is 0.05.}

\item{xaxis.extend.xscale}{The extension factor for the scaling of the x-axis.
Default is NULL (uses extend.xscale).}

\item{xaxis.extend.yscale}{The extension factor for the scaling of the x-axis.
Default is NULL (uses extend.yscale).}

\item{yaxis.extend.xscale}{The extension factor for the scaling of the y-axis.
Default is NULL (uses extend.xscale).}

\item{yaxis.extend.yscale}{The extension factor for the scaling of the y-axis.
Default is NULL (uses extend.yscale).}

\item{name}{The name of the arc sector graphical object. Default is NULL.}

\item{gp}{A graphical parameter object for the arc sector. Default is NULL.}

\item{vp}{A viewport object for the arc sector. Default is NULL.}
}
\value{
A graphical object of class 'grob'.
}
\description{
This function creates a custom arc sector graphical object that can be added to
a plot in R. The arc sector is defined by its starting and ending angles, inner
and outer radii, and center coordinates.
}
\examples{
\dontrun{
# Create an arc sector
arc_sector <- arcSectorGrob(start = 0, end = 180, r0 = 0.5, r1 = 1)
# Add the arc sector to a plot
grid.draw(arc_sector)
}

}
