% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcSegmentsGrob.R
\name{arcSegmentsGrob}
\alias{arcSegmentsGrob}
\title{Create custom arc segments graphical object.}
\usage{
arcSegmentsGrob(
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  start = 0,
  end = 360,
  r0 = 0.5,
  r1 = 1,
  x0 = 0,
  y0 = 0,
  polyline.gp = NULL,
  arrow = NULL,
  n = 100,
  clock.wise = FALSE,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  xscale = NULL,
  yscale = NULL,
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{xmin}{The minimum x value of the segments along the arc. Default is NULL.}

\item{xmax}{The maximum x value of the segments along the arc. Default is NULL.}

\item{ymin}{The minimum y value of the segments along the arc. Default is NULL.}

\item{ymax}{The maximum y value of the segments along the arc. Default is NULL.}

\item{start}{The starting angle of the arc (in degrees). Default is 0.}

\item{end}{The ending angle of the arc (in degrees). Default is 360.}

\item{r0}{The inner radius of the arc. Default is 0.5.}

\item{r1}{The outer radius of the arc. Default is 1.}

\item{x0}{The x-coordinate of the center of the unit circle. Default is 0.}

\item{y0}{The y-coordinate of the center of the unit circle. Default is 0.}

\item{polyline.gp}{A graphical parameter object for the segments' appearance.
Default is NULL.}

\item{arrow}{A list specifying arrow settings for the segments. Default is NULL.}

\item{n}{The number of points used to approximate the segments. Default is 100.}

\item{clock.wise}{A logical value indicating whether the arc is drawn in a
clockwise direction. Default is FALSE.}

\item{extend.xscale}{The extension factor for the scaling of the arc. Default is 0.05.}

\item{extend.yscale}{The extension factor for the scaling of the arc. Default is 0.05.}

\item{xscale}{The scaling range for the x-axis. Default is NULL.}

\item{yscale}{The scaling range for the y-axis. Default is NULL.}

\item{name}{The name of the arc segments graphical object. Default is NULL.}

\item{gp}{A graphical parameter object for the arc segments. Default is NULL.}

\item{vp}{A viewport object for the arc segments. Default is NULL.}
}
\value{
A graphical object of class 'grob'.
}
\description{
This function creates a custom arc segments graphical object that can be added
to a plot in R. The segments are defined by their positions along an arc in
a polar coordinate system.
}
\examples{
\dontrun{
# Create arc segments
arc_segments <- arcSegmentsGrob(xmin = 0, xmax = 45, ymin = 0.5, ymax = 0.7,
                                start = 0, end = 180, r0 = 0.5, r1 = 1)
# Add the arc segments to a plot
grid.draw(arc_segments)
}

}
