% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcTextGrob.R
\name{arcTextGrob}
\alias{arcTextGrob}
\title{Create custom arc text labels graphical object.}
\usage{
arcTextGrob(
  x = NULL,
  y = NULL,
  labels = NULL,
  shift = 0.5,
  start = 0,
  end = 60,
  r0 = 0.5,
  r1 = 1,
  x0 = 0,
  y0 = 0,
  text.gp = NULL,
  nice.facing = TRUE,
  inward = FALSE,
  clock.wise = FALSE,
  curved.label = TRUE,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  xscale = NULL,
  yscale = NULL,
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{x}{The x-coordinates of the text labels. Default is NULL.}

\item{y}{The y-coordinates of the text labels. Default is NULL.}

\item{labels}{The text labels to be displayed. Default is NULL.}

\item{shift}{The factor by which the labels are shifted along the arc. Default is 0.5.}

\item{start}{The starting angle of the arc (in degrees). Default is 0.}

\item{end}{The ending angle of the arc (in degrees). Default is 60.}

\item{r0}{The inner radius of the arc. Default is 0.5.}

\item{r1}{The outer radius of the arc. Default is 1.}

\item{x0}{The x-coordinate of the center of the unit circle. Default is 0.}

\item{y0}{The y-coordinate of the center of the unit circle. Default is 0.}

\item{text.gp}{A graphical parameter object for the text labels' appearance.
Default is NULL.}

\item{nice.facing}{A logical value indicating whether the labels should face in
a "nice" direction. Default is TRUE.}

\item{inward}{A logical value indicating whether the labels should face inward
or outward from the center of the arc. Default is FALSE.}

\item{clock.wise}{A logical value indicating whether the arc is drawn in a
clockwise direction. Default is FALSE.}

\item{curved.label}{A logical value indicating whether the labels should be curved
along the arc. Default is TRUE.}

\item{extend.xscale}{The extension factor for the scaling of the arc. Default is 0.05.}

\item{extend.yscale}{The extension factor for the scaling of the arc. Default is 0.05.}

\item{xscale}{The scaling range for the x-axis. Default is NULL.}

\item{yscale}{The scaling range for the y-axis. Default is NULL.}

\item{name}{The name of the arc text labels graphical object. Default is NULL.}

\item{gp}{A graphical parameter object for the arc text labels. Default is NULL.}

\item{vp}{A viewport object for the arc text labels. Default is NULL.}
}
\value{
A graphical object of class 'grob'.
}
\description{
This function creates a custom arc text labels graphical object that can be added
to a plot in R. The text labels are positioned along an arc in a polar coordinate system.
}
\examples{
\dontrun{
# Create arc text labels
arc_text_labels <- arcTextGrob(x = c(0, 10, 20), y = c(0, 0, 0), labels = c("A", "B", "C"))
# Add the arc text labels to a plot
grid.draw(arc_text_labels)
}

}
