% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-funcs.R
\name{as.radian}
\alias{as.radian}
\title{This function converts degrees to radians.}
\usage{
as.radian(degree)
}
\arguments{
\item{degree}{Numeric vector or scalar containing angles in degrees.}
}
\value{
Numeric vector or scalar with corresponding angles in radians.
}
\description{
This function converts degrees to radians.
}
