% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-funcs.R
\name{as.theta}
\alias{as.theta}
\title{This function converts radians to degrees.}
\usage{
as.theta(radian)
}
\arguments{
\item{radian}{Numeric vector or scalar containing angles in radians.}
}
\value{
Numeric vector or scalar with corresponding angles in degrees.
}
\description{
This function converts radians to degrees.
}
