% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-funcs.R
\name{bedMatTolong}
\alias{bedMatTolong}
\title{Convert a BED matrix to a long format}
\usage{
bedMatTolong(bed.mat = NULL, retain.var = NULL)
}
\arguments{
\item{bed.mat}{A data frame containing BED format data with columns 'chr',
'start', and 'end'.}

\item{retain.var}{A character vector of additional variables to retain in the
long format data frame (optional).}
}
\value{
A long format data frame.
}
\description{
This function converts a BED matrix into a long format data frame for further analysis.
}
\details{
This function orders the input BED matrix by chromosome
and start position, and then creates a long format data frame with columns
for chromosome, start, end, x-coordinate, variable, and y-coordinate.
}
\seealso{
\code{\link{melt}}, \code{\link{order}}
}
