% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_sector.R
\name{facet_sector}
\alias{facet_sector}
\title{Calculate Sector Information for Faceting}
\usage{
facet_sector(
  data = NULL,
  space = c("fixed", "free_x"),
  data.type = c("other", "genomic"),
  sector.gap = 3
)
}
\arguments{
\item{data}{The data frame containing the data (default: NULL).}

\item{space}{Type of space scales for faceting ("fixed" or "free_x") (default: "fixed").}

\item{data.type}{Type of data to use for calculating sectors ("other" or "genomic") (default: "other").}

\item{sector.gap}{Gap between sectors in degrees (default: 3).}
}
\value{
A data frame containing sector information for faceting.
}
\description{
This function calculates sector information for faceting in circular plots.
}
\examples{
\dontrun{
# Example usage of facet_sector
facet_sector(data = my_data, space = "free_x", data.type = "genomic", sector.gap = 5)
}

}
