% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arcarea.R, R/geom_arcdensity.R
\name{geom_arcarea}
\alias{geom_arcarea}
\alias{geom_arcdensity}
\title{Create a stacked area plot in a polar coordinate system.}
\usage{
geom_arcarea(
  mapping = NULL,
  data = NULL,
  stat = "align",
  position = "stack",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  ...,
  polar.every = TRUE,
  outline.type = "upper"
)

geom_arcdensity(
  mapping = NULL,
  data = NULL,
  stat = "density",
  position = "identity",
  ...,
  polar.every = FALSE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  outline.type = "upper"
)
}
\arguments{
\item{mapping}{Aesthetic mapping.}

\item{data}{The data to be plotted.}

\item{stat}{The statistical transformation to be applied (default: "align").}

\item{position}{The position adjustment to be applied (default: "stack").}

\item{na.rm}{Should missing values be removed? (default: FALSE).}

\item{orientation}{Orientation of the area plot (default: NA).}

\item{show.legend}{Should the legend be shown? (default: NA).}

\item{inherit.aes}{Should aesthetics be inherited? (default: TRUE).}

\item{...}{other params.}

\item{polar.every}{Should polar coordinate be used for every data point? (default: TRUE).}

\item{outline.type}{Type of outline (default: "upper").}
}
\value{
A layer to be added to a ggplot object.
}
\description{
\code{geom_arcarea} is used to create a stacked area plot in a polar coordinate system.
}
\seealso{
\code{\link[ggplot2]{geom_area}}
}
