% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arcarrow.R
\name{geom_arcarrow}
\alias{geom_arcarrow}
\title{Create rectangles based on arcs.}
\usage{
geom_arcarrow(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  linejoin = "mitre",
  lineend = "butt",
  width = 0.5,
  start.arrow.len = 6,
  end.arrow.len = 6,
  arrow.start.width = 0.5,
  arrow.end.width = 0.5,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  add.bg = TRUE,
  sector.bg.extend = 0.025,
  add.xaxis = TRUE,
  add.yaxis = TRUE,
  xAxis.params = list(),
  yAxis.params = list(),
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping.}

\item{data}{The data to be plotted.}

\item{stat}{The statistical transformation to be applied (default: "identity").}

\item{position}{The position adjustment to be applied (default: "identity").}

\item{na.rm}{Should missing values be removed? (default: FALSE).}

\item{show.legend}{Should the legend be shown? (default: NA).}

\item{inherit.aes}{Should aesthetics be inherited? (default: TRUE).}

\item{linejoin}{Line join style (default: "mitre").}

\item{lineend}{Line end style (default: "butt").}

\item{width}{Width of the arrow (default: 0.5).}

\item{start.arrow.len}{Length of the starting arrow (default: 6).}

\item{end.arrow.len}{Length of the ending arrow (default: 6).}

\item{arrow.start.width}{Width of the starting arrow (default: 0.5).}

\item{arrow.end.width}{Width of the ending arrow (default: 0.5).}

\item{extend.xscale}{Extension factor for x-axis scale (default: 0.05).}

\item{extend.yscale}{Extension factor for y-axis scale (default: 0.05).}

\item{add.bg}{Should a background be added? (default: TRUE).}

\item{sector.bg.extend}{Extension factor for sector background (default: 0.025).}

\item{add.xaxis}{Should x-axis be added? (default: TRUE).}

\item{add.yaxis}{Should y-axis be added? (default: TRUE).}

\item{xAxis.params}{A list of parameters for x-axis (default: list()).}

\item{yAxis.params}{A list of parameters for y-axis (default: list()).}

\item{...}{Additional parameters to be passed to the \code{GeomArcrect}.}
}
\value{
A layer to be added to a ggplot object.
}
\description{
\code{geom_arcrect} is used to create rectangles based on arcs in a polar plot.
}
\seealso{
\code{\link[ggplot2]{geom_rect}}
}
