% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arcbar.R, R/geom_arccol.R
\name{geom_arcbar}
\alias{geom_arcbar}
\alias{geom_arccol}
\title{Arcbar Geom for Circular Plots}
\usage{
geom_arcbar(
  mapping = NULL,
  data = NULL,
  stat = "bar",
  position = "stack",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  just = 0.5,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  add.bg = TRUE,
  sector.bg.extend = 0.025,
  add.xaxis = TRUE,
  add.yaxis = TRUE,
  xAxis.params = list(),
  yAxis.params = list(),
  ...
)

geom_arccol(
  mapping = NULL,
  data = NULL,
  position = "stack",
  stat = "identity",
  ...,
  just = 0.5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Aesthetic mappings (default: NULL).}

\item{data}{The data frame containing the data (default: NULL).}

\item{stat}{The statistical transformation to use (default: "bar").}

\item{position}{The position adjustment for overlapping objects (default: "stack").}

\item{na.rm}{A logical value indicating whether missing values should be removed (default: FALSE).}

\item{show.legend}{A logical value indicating whether to show the legend (default: NA).}

\item{inherit.aes}{A logical value indicating whether to inherit aesthetics from the plot (default: TRUE).}

\item{just}{A numeric value (between 0 and 1) controlling the justification of the arc bars (default: 0.5).}

\item{extend.xscale}{The extension factor for the x-axis scale (default: 0.05).}

\item{extend.yscale}{The extension factor for the y-axis scale (default: 0.05).}

\item{add.bg}{A logical value indicating whether to add a background (default: TRUE).}

\item{sector.bg.extend}{The extension factor for the sector background (default: 0.025).}

\item{add.xaxis}{A logical value indicating whether to add the x-axis (default: TRUE).}

\item{add.yaxis}{A logical value indicating whether to add the y-axis (default: TRUE).}

\item{xAxis.params}{A list of parameters for customizing the x-axis (default: list()).}

\item{yAxis.params}{A list of parameters for customizing the y-axis (default: list()).}

\item{...}{other params}
}
\value{
A circular plot with arc-shaped bars.
}
\description{
This function adds arc-shaped bars to a circular plot created with ggplot2,
allowing you to represent data with radial bars.
}
