% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arcboxplot.R
\name{geom_arcboxplot}
\alias{geom_arcboxplot}
\title{Create an arc boxplot layer for ggplot2.}
\usage{
geom_arcboxplot(
  mapping = NULL,
  data = NULL,
  stat = "boxplot",
  position = "dodge2",
  outliers = TRUE,
  outlier.colour = NULL,
  outlier.color = NULL,
  outlier.fill = NULL,
  outlier.shape = 19,
  outlier.size = 1.5,
  outlier.stroke = 0.5,
  outlier.alpha = NULL,
  notch = FALSE,
  notchwidth = 0.5,
  staplewidth = 0,
  varwidth = FALSE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  add.bg = TRUE,
  sector.bg.extend = 0.025,
  add.xaxis = TRUE,
  add.yaxis = TRUE,
  xAxis.params = list(),
  yAxis.params = list(),
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping.}

\item{data}{The data to be plotted.}

\item{stat}{The statistical transformation to be applied (default: "arcboxplot").}

\item{position}{The position adjustment to be applied (default: "dodge2").}

\item{outliers}{Should outliers be displayed? (default: TRUE)}

\item{outlier.colour}{Color of the outlier points (default: NULL).}

\item{outlier.color}{Color of the outlier points (default: NULL).}

\item{outlier.fill}{Fill color of the outlier points (default: NULL).}

\item{outlier.shape}{Shape of the outlier points (default: 19).}

\item{outlier.size}{Size of the outlier points (default: 1.5).}

\item{outlier.stroke}{Stroke width of the outlier points (default: 0.5).}

\item{outlier.alpha}{Alpha transparency of the outlier points (default: NULL).}

\item{notch}{Should notches be added to the boxplots? (default: FALSE)}

\item{notchwidth}{Width of notches (default: 0.5).}

\item{staplewidth}{Width of staple lines (default: 0).}

\item{varwidth}{Should box widths vary with sample sizes? (default: FALSE)}

\item{na.rm}{Should missing values be removed? (default: FALSE)}

\item{orientation}{Orientation of the arc boxplot (default: NA).}

\item{show.legend}{Should the legend be shown? (default: NA).}

\item{inherit.aes}{Should aesthetics be inherited? (default: TRUE)}

\item{extend.xscale}{Extension factor for x-axis scale (default: 0.05).}

\item{extend.yscale}{Extension factor for y-axis scale (default: 0.05).}

\item{add.bg}{Should a background be added? (default: TRUE)}

\item{sector.bg.extend}{Extension factor for sector background (default: 0.025).}

\item{add.xaxis}{Should x-axis be added? (default: TRUE)}

\item{add.yaxis}{Should y-axis be added? (default: TRUE)}

\item{xAxis.params}{A list of parameters for x-axis (default: list()).}

\item{yAxis.params}{A list of parameters for y-axis (default: list()).}

\item{...}{Additional parameters to be passed to the \code{GeomArcboxplot}.}
}
\value{
A layer to be added to a ggplot object.
}
\description{
\code{geom_arcboxplot} is used to create an arc boxplot layer for ggplot2.
}
\seealso{
\code{\link[ggplot2]{geom_boxplot}}
}
