% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arcline.R
\name{geom_arcline}
\alias{geom_arcline}
\title{Create a line segment layer with options for arrowheads and line style.}
\usage{
geom_arcline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  arrow = NULL,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  add.bg = TRUE,
  sector.bg.extend = 0.025,
  add.xaxis = TRUE,
  add.yaxis = TRUE,
  xAxis.params = list(),
  yAxis.params = list(),
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping.}

\item{data}{The data to be plotted.}

\item{stat}{The statistical transformation to be applied (default: "identity").}

\item{position}{The position adjustment to be applied (default: "identity").}

\item{na.rm}{Should missing values be removed? (default: FALSE)}

\item{show.legend}{Should the legend be shown? (default: NA).}

\item{inherit.aes}{Should aesthetics be inherited? (default: TRUE)}

\item{arrow}{A list of arrow properties (default: NULL).}

\item{lineend}{Line end style (default: "butt").}

\item{linejoin}{Line join style (default: "round").}

\item{linemitre}{Line mitre limit (default: 10).}

\item{extend.xscale}{Extension factor for x-axis scale (default: 0.05).}

\item{extend.yscale}{Extension factor for y-axis scale (default: 0.05).}

\item{add.bg}{Should a background be added? (default: TRUE)}

\item{sector.bg.extend}{Extension factor for sector background (default: 0.025).}

\item{add.xaxis}{Should x-axis be added? (default: TRUE)}

\item{add.yaxis}{Should y-axis be added? (default: TRUE)}

\item{xAxis.params}{A list of parameters for x-axis (default: list()).}

\item{yAxis.params}{A list of parameters for y-axis (default: list()).}

\item{...}{Additional parameters to be passed to the \code{GeomArcline}.}
}
\value{
A layer to be added to a ggplot object.
}
\description{
\code{geom_arcline} is used to create a line segment layer with options for arrowheads and line style.
}
\seealso{
\code{\link[ggcirclize]{arcLinesGrob}}
}
