% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arclinerange.R
\name{geom_arclinerange}
\alias{geom_arclinerange}
\title{Create a range of line segments with options for line style and orientation.}
\usage{
geom_arclinerange(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  lineend = "butt",
  linejoin = "round",
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  add.bg = TRUE,
  sector.bg.extend = 0.025,
  add.xaxis = TRUE,
  add.yaxis = TRUE,
  xAxis.params = list(),
  yAxis.params = list(),
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping.}

\item{data}{The data to be plotted.}

\item{stat}{The statistical transformation to be applied (default: "identity").}

\item{position}{The position adjustment to be applied (default: "identity").}

\item{na.rm}{Should missing values be removed? (default: FALSE)}

\item{orientation}{Orientation of the line segments (default: NA).}

\item{show.legend}{Should the legend be shown? (default: NA).}

\item{inherit.aes}{Should aesthetics be inherited? (default: TRUE)}

\item{lineend}{Line end style (default: "butt").}

\item{linejoin}{Line join style (default: "round").}

\item{extend.xscale}{Extension factor for x-axis scale (default: 0.05).}

\item{extend.yscale}{Extension factor for y-axis scale (default: 0.05).}

\item{add.bg}{Should a background be added? (default: TRUE)}

\item{sector.bg.extend}{Extension factor for sector background (default: 0.025).}

\item{add.xaxis}{Should x-axis be added? (default: TRUE)}

\item{add.yaxis}{Should y-axis be added? (default: TRUE)}

\item{xAxis.params}{A list of parameters for x-axis (default: list()).}

\item{yAxis.params}{A list of parameters for y-axis (default: list()).}

\item{...}{Additional parameters to be passed to the \code{GeomArclinerange}.}
}
\value{
A layer to be added to a ggplot object.
}
\description{
\code{geom_arclinerange} is used to create a range of line segments with options for line style and orientation.
}
\seealso{
\code{\link[ggplot2]{geom_linerange}}
}
