% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arcsector.R
\name{geom_arcsector}
\alias{geom_arcsector}
\title{Create sectors based on arcs.}
\usage{
geom_arcsector(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  add.bg = TRUE,
  sector.bg.extend = 0.025,
  add.xaxis = FALSE,
  add.yaxis = FALSE,
  xAxis.params = list(),
  yAxis.params = list(),
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping.}

\item{data}{The data to be plotted.}

\item{stat}{The statistical transformation to be applied (default: "identity").}

\item{position}{The position adjustment to be applied (default: "identity").}

\item{na.rm}{Should missing values be removed? (default: FALSE).}

\item{show.legend}{Should the legend be shown? (default: NA).}

\item{inherit.aes}{Should aesthetics be inherited? (default: TRUE).}

\item{extend.xscale}{Extension factor for x-axis scale (default: 0.05).}

\item{extend.yscale}{Extension factor for y-axis scale (default: 0.05).}

\item{add.bg}{Should a background be added? (default: TRUE).}

\item{sector.bg.extend}{Extension factor for sector background (default: 0.025).}

\item{add.xaxis}{Should x-axis be added? (default: FALSE).}

\item{add.yaxis}{Should y-axis be added? (default: FALSE).}

\item{xAxis.params}{A list of parameters for x-axis (default: list()).}

\item{yAxis.params}{A list of parameters for y-axis (default: list()).}

\item{...}{Additional parameters to be passed to the \code{GeomArcsector}.}
}
\value{
A layer to be added to a ggplot object.
}
\description{
\code{geom_arcsector} is used to create sectors based on arcs in a polar plot.
}
\seealso{
\code{\link[ggplot2]{geom_polygon}}
}
