% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arctext.R
\name{geom_arctext}
\alias{geom_arctext}
\title{Create text labels along arcs.}
\usage{
geom_arctext(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  curved.label = TRUE,
  shift = 0.65,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  add.bg = TRUE,
  sector.bg.extend = 0.025,
  add.xaxis = TRUE,
  add.yaxis = TRUE,
  xAxis.params = list(),
  yAxis.params = list(),
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping.}

\item{data}{The data to be plotted.}

\item{stat}{The statistical transformation to be applied (default: "identity").}

\item{position}{The position adjustment to be applied (default: "identity").}

\item{na.rm}{Should missing values be removed? (default: FALSE).}

\item{show.legend}{Should the legend be shown? (default: NA).}

\item{inherit.aes}{Should aesthetics be inherited? (default: TRUE).}

\item{curved.label}{Should labels be curved along the arc? (default: TRUE).}

\item{shift}{Shift factor for label placement along the arc (default: 0.65).}

\item{extend.xscale}{Extension factor for x-axis scale (default: 0.05).}

\item{extend.yscale}{Extension factor for y-axis scale (default: 0.05).}

\item{add.bg}{Should a background be added? (default: TRUE).}

\item{sector.bg.extend}{Extension factor for sector background (default: 0.025).}

\item{add.xaxis}{Should x-axis be added? (default: TRUE).}

\item{add.yaxis}{Should y-axis be added? (default: TRUE).}

\item{xAxis.params}{A list of parameters for x-axis (default: list()).}

\item{yAxis.params}{A list of parameters for y-axis (default: list()).}

\item{...}{Additional parameters to be passed to the \code{GeomArctext}.}
}
\value{
A layer to be added to a ggplot object.
}
\description{
\code{geom_arctext} is used to create text labels along arcs in a polar plot.
}
\seealso{
\code{\link[ggplot2]{geom_text}}
}
