% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arcviolin.R
\name{geom_arcviolin}
\alias{geom_arcviolin}
\title{Create a violin-like arc density plot in a polar coordinate system.}
\usage{
geom_arcviolin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "dodge",
  draw_quantiles = NULL,
  trim = TRUE,
  bounds = c(-Inf, Inf),
  scale = "area",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  polar.every = FALSE,
  quantiles.colour = "black",
  quantiles.linewidth = 1,
  quantiles.linetype = 1,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  add.bg = TRUE,
  sector.bg.extend = 0.025,
  add.xaxis = TRUE,
  add.yaxis = TRUE,
  xAxis.params = list(),
  yAxis.params = list(),
  lineend = "butt",
  linejoin = "round",
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping.}

\item{data}{The data to be plotted.}

\item{stat}{The statistical transformation to be applied (default: "arcydensity").}

\item{position}{The position adjustment to be applied (default: "dodge").}

\item{draw_quantiles}{Should quantiles be drawn? (default: NULL).}

\item{trim}{Should the violin be trimmed? (default: TRUE).}

\item{bounds}{The bounds for the density calculation (default: c(-Inf, Inf)).}

\item{scale}{The scaling method (default: "area").}

\item{na.rm}{Should missing values be removed? (default: FALSE).}

\item{orientation}{Orientation of the violin plot (default: NA).}

\item{show.legend}{Should the legend be shown? (default: NA).}

\item{inherit.aes}{Should aesthetics be inherited? (default: TRUE).}

\item{polar.every}{Should polar coordinate be used for every data point? (default: FALSE).}

\item{quantiles.colour}{Colour of quantile lines (default: "black").}

\item{quantiles.linewidth}{Line width of quantile lines (default: 1).}

\item{quantiles.linetype}{Line type of quantile lines (default: 1).}

\item{extend.xscale}{Extension factor for x-axis scale (default: 0.05).}

\item{extend.yscale}{Extension factor for y-axis scale (default: 0.05).}

\item{add.bg}{Should a background be added? (default: TRUE).}

\item{sector.bg.extend}{Extension factor for sector background (default: 0.025).}

\item{add.xaxis}{Should x-axis be added? (default: TRUE).}

\item{add.yaxis}{Should y-axis be added? (default: TRUE).}

\item{xAxis.params}{A list of parameters for x-axis (default: list()).}

\item{yAxis.params}{A list of parameters for y-axis (default: list()).}

\item{lineend}{Type of line end (default: "butt").}

\item{linejoin}{Type of line join (default: "round").}

\item{...}{Additional parameters to be passed to the \code{GeomArcviolin}.}
}
\value{
A layer to be added to a ggplot object.
}
\description{
\code{geom_arcviolin} is used to create a violin-like arc density plot in a polar coordinate system.
}
\seealso{
\code{\link[ggplot2]{geom_violin}}
}
