% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_chordDiagram.R
\name{geom_chordDiagram}
\alias{geom_chordDiagram}
\title{Create a chord diagram plot.}
\usage{
geom_chordDiagram(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  start.arrow = FALSE,
  end.arrow = FALSE,
  curve.arrow = NULL,
  arrow.len = 0.01,
  link.sector.space = 0.05,
  sector.height = 0.1,
  nice.facing = TRUE,
  sector.fill = NULL,
  sector.order = NULL,
  vp.angle = NULL,
  link.sort = TRUE,
  link.decreasing = c(TRUE, TRUE),
  directional = 0,
  diffHeight = 0,
  curve.height.low = 0.6,
  link2line = FALSE,
  big.gap = 15,
  strip.label = TRUE,
  strip.label.space = 0.15,
  strip.label.fontface = "bold",
  strip.label.col = "black",
  strip.label.size = 10,
  space = c("free_x", "fixed"),
  sector.gap = 3,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  add.bg = TRUE,
  sector.bg.extend = 0.025,
  add.xaxis = TRUE,
  add.yaxis = FALSE,
  xAxis.params = list(),
  yAxis.params = list(),
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping.}

\item{data}{The data to be plotted.}

\item{stat}{The statistical transformation to be applied (default is "identity").}

\item{position}{The position adjustment method to be used (default is "identity").}

\item{na.rm}{Logical value indicating whether to remove NA values (default is FALSE).}

\item{show.legend}{Logical value indicating whether to display the legend (default is NA).}

\item{inherit.aes}{Logical value indicating whether to inherit aesthetics (default is TRUE).}

\item{start.arrow}{Logical value indicating whether to include arrows at the start of links (default is FALSE).}

\item{end.arrow}{Logical value indicating whether to include arrows at the end of links (default is FALSE).}

\item{curve.arrow}{The curvature of the arrows (default is NULL).}

\item{arrow.len}{The length of arrows (default is 0.01).}

\item{link.sector.space}{The space between links and sectors (default is 0.05).}

\item{sector.height}{The height of sectors (default is 0.1).}

\item{nice.facing}{Logical value indicating whether to use nice facing for sectors (default is TRUE).}

\item{sector.fill}{Fill color for sectors (default is NULL).}

\item{sector.order}{Order of sectors (default is NULL).}

\item{vp.angle}{Angle of the viewing perspective (default is NULL).}

\item{link.sort}{Logical value indicating whether to sort links (default is TRUE).}

\item{link.decreasing}{Logical vector indicating whether links should be decreasing in size (default is c(TRUE, TRUE)).}

\item{directional}{Numeric value indicating directionality (default is 0).}

\item{diffHeight}{Numeric value indicating height difference (default is 0).}

\item{curve.height.low}{Numeric value indicating the low point of the curve (default is 0.6).}

\item{link2line}{Logical value indicating whether to convert links to lines (default is FALSE).}

\item{big.gap}{Numeric value indicating the size of a big gap (default is 15).}

\item{strip.label}{Logical value indicating whether to display strip labels (default is TRUE).}

\item{strip.label.space}{Numeric value indicating the space for strip labels (default is 0.15).}

\item{strip.label.fontface}{Font face for strip labels (default is "bold").}

\item{strip.label.col}{Color for strip labels (default is "black").}

\item{strip.label.size}{Size of strip labels (default is 10).}

\item{space}{Space setting for x and y scales (default is c("free_x", "fixed")).}

\item{sector.gap}{Numeric value indicating the gap between sectors (default is 3).}

\item{extend.xscale}{Numeric value indicating the extension of the x-scale (default is 0.05).}

\item{extend.yscale}{Numeric value indicating the extension of the y-scale (default is 0.05).}

\item{add.bg}{Logical value indicating whether to add a background (default is TRUE).}

\item{sector.bg.extend}{Numeric value indicating the extension of the sector background (default is 0.025).}

\item{add.xaxis}{Logical value indicating whether to add an x-axis (default is TRUE).}

\item{add.yaxis}{Logical value indicating whether to add a y-axis (default is FALSE).}

\item{xAxis.params}{List of parameters for the x-axis (default is list()).}

\item{yAxis.params}{List of parameters for the y-axis (default is list()).}

\item{...}{Additional parameters.}
}
\value{
A chord diagram plot.
}
\description{
This function creates a chord diagram plot using the specified data and
mapping.
}
