% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cpoint.R
\name{geom_cpoint}
\alias{geom_cpoint}
\title{Add Custom Points to a ggplot2 Plot}
\usage{
geom_cpoint(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Aesthetic mappings (default: NULL).}

\item{data}{The data frame containing the data (default: NULL).}

\item{stat}{The statistical transformation to use (default: "identity").}

\item{position}{The position adjustment for overlapping points (default: "identity").}

\item{...}{Additional parameters to customize the appearance of the points.}

\item{na.rm}{A logical value indicating whether to remove missing values (default: FALSE).}

\item{show.legend}{A logical value indicating whether to show the legend (default: NA).}

\item{inherit.aes}{A logical value indicating whether to inherit aesthetics from the plot (default: TRUE).}
}
\value{
A layer to be added to a ggplot object.
}
\description{
This function allows you to add custom points to a ggplot2 plot. It is a
geom layer for creating scatterplots
}
