% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_genomicNestedZoom.R
\name{geom_genomicNestedZoom}
\alias{geom_genomicNestedZoom}
\title{Create a genomic nested zoom plot}
\usage{
geom_genomicNestedZoom(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  chrom_data = NULL,
  keep.all.chrom = FALSE,
  zoom_data = NULL,
  geom = c("point", "line", "rect", "tile", "area", "rainfall", "density"),
  extend.start = 0,
  extend.end = 0,
  nested.r = 0.02,
  panel.gap = 0.02,
  zoom.panel.height = NULL,
  sub.zoom.gap = 1,
  zoom_pos = c("bottom", "top"),
  n = 100,
  nested_fill = NULL,
  nested_col = "black",
  nested_lty = "solid",
  nested_linewidth = 0.5,
  nested_alpha = c(0.2, 0.5),
  zoom.free.sec = FALSE,
  zoom.panel.extend.xscale = 0.05,
  mode = c("min", "max", "mean", "left", "right"),
  normalize_to_width = FALSE,
  window.size = 1e+07,
  count_by = c("percent", "number"),
  polar.every = FALSE,
  space = c("free_x", "fixed"),
  sector.gap = 3,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping.}

\item{data}{The data to be plotted.}

\item{stat}{The statistical transformation to apply to the data. Default is "identity".}

\item{position}{The position adjustment to be applied to the data. Default is "identity".}

\item{na.rm}{Should missing values be removed? Default is FALSE.}

\item{show.legend}{Should a legend be shown? Default is NA.}

\item{inherit.aes}{Should aesthetics be inherited? Default is TRUE.}

\item{chrom_data}{Data containing chromosome information.}

\item{keep.all.chrom}{Should all chromosomes be kept? Default is FALSE.}

\item{zoom_data}{Data for zoomed-in regions which includes 3 columns("chr","start","end").}

\item{geom}{Type of geom to be used. Default is "point".}

\item{extend.start}{Extend the start of the zoom panel. Default is 0.}

\item{extend.end}{Extend the end of the zoom panel. Default is 0.}

\item{nested.r}{Radius of nested genomic regions. Default is 0.02.}

\item{panel.gap}{Gap between zoom panels. Default is 0.02.}

\item{zoom.panel.height}{Height of the zoom panel. Default is NULL.}

\item{sub.zoom.gap}{Gap between sub-zoom panels. Default is 1.}

\item{zoom_pos}{Position of zoom panels (bottom or top). Default is c("bottom", "top").}

\item{n}{Number of points for create plot. Default is 100.}

\item{nested_fill}{Fill color for nested regions. Default is NULL.}

\item{nested_col}{Color for nested regions' outline. Default is "black".}

\item{nested_lty}{Line type for nested regions' outline. Default is "solid".}

\item{nested_linewidth}{Line width for nested regions' outline. Default is 0.5.}

\item{nested_alpha}{Alpha transparency for nested regions. Default is c(0.2, 0.5).}

\item{zoom.free.sec}{Allow free scaling for zoomed-in sections. Default is FALSE.}

\item{zoom.panel.extend.xscale}{Extend the x-scale for zoom panels. Default is 0.05.}

\item{mode}{Details see circlize::rainfallTransform.}

\item{normalize_to_width}{Details see circlize::rainfallTransform.}

\item{window.size}{Size of the genomic window. Default is 1e7.
Details see circlize::genomicDensity.}

\item{count_by}{Method for counting data (percent or number). Details see circlize::genomicDensity.}

\item{polar.every}{Display sectors at fixed angles (for circular plots). Default is FALSE.}

\item{space}{Space type for panels (free_x or fixed). Default is c("free_x", "fixed").}

\item{sector.gap}{Gap between sectors (for circular plots). Default is 3.}

\item{...}{Additional arguments to be passed to the underlying layer function.}
}
\value{
A ggplot2 plot object.
}
\description{
This function creates a genomic nested zoom plot, allowing you to visualize
nested genomic regions and zoom in on specific sections of the data. It can be
used with various geoms such as points, lines, rectangles, areas, rainfall,
and density.
}
