% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_trackgenomicline.R
\name{geom_trackgenomicline}
\alias{geom_trackgenomicline}
\title{Track Genomic Line Plot}
\usage{
geom_trackgenomicline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  chrom_data = NULL,
  keep.all.chrom = FALSE,
  strip.label = TRUE,
  strip.label.pos = c("top", "bottom"),
  strip.label.space = 0.15,
  strip.label.fontface = "bold",
  strip.label.col = "black",
  strip.label.size = 10,
  scales = c("fixed", "free", "free_x", "free_y"),
  space = c("free_x", "fixed"),
  sector.gap = 3,
  extend.xscale = 0.05,
  extend.yscale = 0.05,
  add.bg = TRUE,
  sector.bg.extend = 0.025,
  add.xaxis = TRUE,
  add.yaxis = FALSE,
  xAxis.params = list(),
  yAxis.params = list(),
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping (required).}

\item{data}{The data frame containing the data (required).}

\item{stat}{The statistical transformation to apply (default: "identity").}

\item{position}{The position adjustment method (default: "identity").}

\item{na.rm}{A logical value indicating whether to remove NA values (default: FALSE).}

\item{show.legend}{A logical value indicating whether to show the legend (default: NA).}

\item{inherit.aes}{A logical value indicating whether to inherit aesthetics from the parent plot (default: TRUE).}

\item{chrom_data}{Data frame with 2 columns containing chromosome information (default: NULL).}

\item{keep.all.chrom}{A logical value indicating whether to keep all chromosomes in the plot (default: FALSE).}

\item{strip.label}{A logical value indicating whether to add strip labels (default: TRUE).}

\item{strip.label.pos}{Position of strip labels ("top" or "bottom") (default: c("top", "bottom")).}

\item{strip.label.space}{Space between strip labels (default: 0.15).}

\item{strip.label.fontface}{Fontface for strip labels (default: "bold").}

\item{strip.label.col}{Color of strip labels (default: "black").}

\item{strip.label.size}{Size of strip labels (default: 10).}

\item{scales}{Type of scales for x and y axes ("fixed", "free", "free_x", "free_y")
(default: c("fixed","free","free_x","free_y"): unused arg now).}

\item{space}{Type of space scales ("free_x", "fixed") (default: c("free_x","fixed")).}

\item{sector.gap}{Gap between sectors in polar coordinates (default: 3).}

\item{extend.xscale}{Extension of x-axis scale (default: 0.05).}

\item{extend.yscale}{Extension of y-axis scale (default: 0.05).}

\item{add.bg}{A logical value indicating whether to add background to the plot (default: TRUE).}

\item{sector.bg.extend}{Extension of sector background (default: 0.025).}

\item{add.xaxis}{A logical value indicating whether to add x-axis (default: TRUE).}

\item{add.yaxis}{A logical value indicating whether to add y-axis (default: FALSE).}

\item{xAxis.params}{A list of parameters for customizing the x-axis (default: list()).}

\item{yAxis.params}{A list of parameters for customizing the y-axis (default: list()).}

\item{...}{Additional arguments to be passed to the underlying layer function.}
}
\value{
A ggplot2 layer.
}
\description{
This function creates a genomic line plot using track-based visualization.
}
