% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-funcs.R
\name{get_chrom_data}
\alias{get_chrom_data}
\title{Fetch chromosome size from ggcirclize package}
\usage{
get_chrom_data(genome = "hg19")
}
\arguments{
\item{genome}{Genome version includes hg38,hg19,mm39,mm10 or mm9.}
}
\value{
A 2-column data.frame includes chromosome length.
}
\description{
Fetch chromosome size from ggcirclize package
}
