% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-funcs.R
\name{get_fixedHeight}
\alias{get_fixedHeight}
\title{Calculate Height for Fixed Arc Angle Span}
\usage{
get_fixedHeight(start, end, r0, r1, height)
}
\arguments{
\item{start}{Starting angle in degrees.}

\item{end}{Ending angle in degrees.}

\item{r0}{Inner radius.}

\item{r1}{Outer radius.}

\item{height}{Fixed height (default: NULL).}
}
\value{
The calculated height for the specified arc angle span.
}
\description{
This function calculates the height for a fixed arc angle span based on the specified
starting angle, ending angle, inner radius, outer radius, and height.
}
