% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-funcs.R
\name{get_height}
\alias{get_height}
\title{Calculate Height Between Arc Angles}
\usage{
get_height(start, end, r0, r1)
}
\arguments{
\item{start}{Starting angle in degrees.}

\item{end}{Ending angle in degrees.}

\item{r0}{Inner radius.}

\item{r1}{Outer radius.}
}
\value{
The height between the specified arc angles.
}
\description{
This function calculates the height between two arc angles based on the specified
starting angle, ending angle, inner radius, and outer radius.
}
