% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-funcs.R
\name{get_symmetric_val}
\alias{get_symmetric_val}
\title{Get Symmetric Values from a Matrix}
\usage{
get_symmetric_val(mat = NULL, keep.diag = TRUE)
}
\arguments{
\item{mat}{A matrix from which symmetric values will be extracted.}

\item{keep.diag}{Logical value indicating whether to keep the diagonal elements (default is TRUE).}
}
\value{
A data frame with three columns: 'from', 'to', and 'value', representing the row index, column index, and
corresponding value of the symmetric elements.
}
\description{
This function extracts the symmetric values from a matrix and returns them in a long format.
}
