% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcirclize-package.R, R/ggcirclize.R
\docType{package}
\name{ggcirclize}
\alias{ggcirclize}
\alias{package-ggcirclize}
\title{visualizing data on circular layout based on grammar of graphics}
\usage{
ggcirclize(
  data = NULL,
  mapping = aes(),
  mg.t = 2,
  mg.r = 1,
  mg.b = 2,
  mg.l = 1,
  environment = parent.frame(),
  ...
)
}
\arguments{
\item{data}{The data frame containing the data (default: NULL).}

\item{mapping}{Aesthetic mapping (default: aes(start = 0, end = 180, r0 = 0.5, r1 = 1, clock.wise = FALSE)).}

\item{mg.t}{Top margin for the plot (default: 2).}

\item{mg.r}{Right margin for the plot (default: 1).}

\item{mg.b}{Bottom margin for the plot (default: 2).}

\item{mg.l}{Left margin for the plot (default: 1).}

\item{environment}{The environment to use for evaluating the function (default: parent.frame()).}

\item{...}{Additional arguments to be passed to ggplot function.}
}
\value{
A circular plot created with ggplot2.
}
\description{
This function creates a circular plot using ggplot2 with specified data and aesthetics.
}
\examples{
\dontrun{
# Example usage of ggcirclize
ggcirclize(data = my_data, mapping = aes(x = value, y = category))
}

}
\author{
Jun Zhang
}
