% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromInfo-Data.R
\docType{data}
\name{hg38_chrom_info}
\alias{hg38_chrom_info}
\title{This is a test data for ggcirclize package}
\format{
An object of class \code{list} of length 2.
}
\usage{
hg38_chrom_info
}
\description{
This is a test data for ggcirclize package
}
\author{
Jun Zhang
}
\keyword{datasets}
