% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartLabelAlignGrob.R
\name{smartLabelAlignGrob}
\alias{smartLabelAlignGrob}
\title{Smart Label Align Grob}
\usage{
smartLabelAlignGrob(
  all.label = NULL,
  mark.label = NULL,
  x = 0.5,
  y = 0.5,
  use.smartAlign2 = FALSE,
  link.line.length = 0.025,
  link.label.space = 0.02,
  mark.scale = c(0.01, 0.99),
  link.label.gp = gpar(fontsize = 10),
  link.line.gp = gpar(fill = "black", col = "black"),
  link.circle.start.gp = gpar(fill = "black", col = "black"),
  link.circle.end.gp = gpar(fill = "black", col = "black"),
  link.start.type = c("line", "circle", "arrow"),
  link.end.type = c("line", "circle", "arrow"),
  circle.arrow.size = c(0.01, 0.01),
  pos = c("right", "left", "top", "bottom"),
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{all.label}{A vector of all labels.}

\item{mark.label}{The label to be aligned.}

\item{x}{The x-coordinate for the alignment.}

\item{y}{The y-coordinate for the alignment.}

\item{use.smartAlign2}{Logical, whether to use smart alignment.}

\item{link.line.length}{Length of the connecting lines.}

\item{link.label.space}{Spacing between the label and connecting lines.}

\item{mark.scale}{A numeric vector specifying the scale for mark positioning.}

\item{link.label.gp}{A graphical parameter for the label appearance.}

\item{link.line.gp}{A graphical parameter for the connecting lines.}

\item{link.circle.start.gp}{A graphical parameter for the start circle (if used).}

\item{link.circle.end.gp}{A graphical parameter for the end circle (if used).}

\item{link.start.type}{Type of shape at the start of the connecting lines ("line", "circle", or "arrow").}

\item{link.end.type}{Type of shape at the end of the connecting lines ("line", "circle", or "arrow").}

\item{circle.arrow.size}{A numeric vector specifying the size of circles or arrows.}

\item{pos}{Alignment position ("right", "left", "top", or "bottom").}

\item{name}{A name for the graphical object.}

\item{gp}{A graphical parameter for the entire grob.}

\item{vp}{A viewport for the grob.}
}
\value{
A graphical object for smart label alignment.
}
\description{
This function creates a graphical object that aligns a labeled mark with a set of labels on a plot.
}
